requirements.txt
apache-airflow==2.6.3
apache-airflow-providers-common-sql==1.6.0
apache-airflow-providers-ftp==3.4.2
apache-airflow-providers-google==10.4.0
apache-airflow-providers-http==4.5.0
apache-airflow-providers-imap==3.2.2
apache-airflow-providers-sqlite==3.4.2
pytest=7.4.2
future=0.18.3
